requirements.txt
Django==2.0.2
django-bootstrap-breadcrumbs==0.9.0
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-webpack-loader==0.5.0
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
typing==3.6.4
Werkzeug==0.14.1
